+++ lib/dotup_dart_logger.dart
@@ -1,11 +1,13 @@
 library dotup_dart_logger;
 
+export 'src/formater/AnsiEscape.dart';
 export 'src/formater/PipeSeparatedLogFormater.dart';
+export 'src/formater/PrettyFormater.dart';
 export 'src/writer/ConsoleLogWriter.dart';
-export 'src/ILogFormater.dart';
-export 'src/ILogWriter.dart';
-export 'src/ILogger.dart';
+export 'src/interfaces/ILogFormater.dart';
+export 'src/interfaces/ILogger.dart';
+export 'src/interfaces/ILogWriter.dart';
 export 'src/LogEntry.dart';
+export 'src/Logger.dart';
 export 'src/LogLevel.dart';
 export 'src/LogManager.dart';
-export 'src/Logger.dart';