+++ lib/src/ILogFormater.dart
@@ -1,5 +1,0 @@
-import 'LogEntry.dart';
-
-abstract class ILogFormater {
-  String format(LogEntry logEntry);
-}