+++ lib/src/Utils.dart
@@ -1,0 +1,15 @@
+import 'LogEntry.dart';
+
+// ignore: prefer_generic_function_type_aliases
+typedef void LogHandler(LogEntry logEntry);
+
+extension IterableExtension<T> on Iterable<T> {
+  T? firstWhereOrNull(bool Function(T element) test) {
+    for (var element in this) {
+      if (test(element)) {
+        return element;
+      }
+    }
+    return null;
+  }
+}