+++ lib/src/interfaces/ILogger.dart
@@ -1,8 +1,8 @@
 abstract class ILogger {
   void console(Object message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
-  void debug(String message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
-  void info(String message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
-  void warn(String message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
+  void debug(Object message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
+  void info(Object message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
+  void warn(Object message, {Map<String, dynamic>? data, Object? error, StackTrace? stackTrace, String? source});
   void error(Error error, {Map<String, dynamic>? data, StackTrace? stackTrace, String? source});
   void exception(Exception error, {Map<String, dynamic>? data, StackTrace? stackTrace, String? source});
 }